Update User model and Stripe controller, fix revenue trend display, and add login alert and sweetalert2 functionality

+++ frontend/src/components/HistoryCard.tsx
@@ -6,6 +6,7 @@ import DeleteReservation from "@/libs/DeleteReservation";
 import { faGear, faPenToSquare, faTrash, faX, faXmark } from "@fortawesome/free-solid-svg-icons"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { useState } from "react";
+import Swal from "sweetalert2";
 
 
 
@@ -21,8 +22,21 @@ export default function HistoryCard( {reservation} : {reservation : Reservation}
     
     const handleDeleteButton = () => {
         if(currentUser && rid){
-            DeleteReservation(currentUser.token, rid)
-            location.reload()
+            Swal.fire({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!'
+              }).then((result) => {
+                if (result.isConfirmed) {
+                    DeleteReservation(currentUser.token, rid)
+                }
+              })
+            
+            
         }
     }
